+++ cmd/metrics/http.go
@@ -1,0 +1,66 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"github.com/gin-gonic/gin"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/sirupsen/logrus"
+	"golang.org/x/sync/errgroup"
+	"net/http"
+	"time"
+)
+
+func runHttpServer(
+	ctx context.Context,
+	listenAddress string,
+	log logrus.FieldLogger,
+) error {
+	router := gin.Default()
+	router.GET("/ping", pong())
+	router.GET("/metrics", metrics())
+
+	srv := &http.Server{
+		Addr:    listenAddress,
+		Handler: router,
+	}
+
+	wg, ctx := errgroup.WithContext(ctx)
+	wg.Go(func() error {
+		<-ctx.Done()
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		log.Infof("HTTP server shutting down...")
+		if err := srv.Shutdown(ctx); err != nil && !errors.Is(err, context.Canceled) {
+			log.WithError(err).Infof("HTTP server shutdown failed")
+			return err
+		}
+		return nil
+	})
+
+	wg.Go(func() error {
+		log.Infof("HTTP server accepting requests on %q", listenAddress)
+		if err := srv.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
+			log.WithError(err).Infof("unexpected error from HTTP server")
+			return err
+		}
+		log.Infof("HTTP server finished, terminating...")
+		return nil
+	})
+	return wg.Wait()
+}
+
+func pong() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.JSON(200, gin.H{
+			"message": "pong",
+		})
+	}
+}
+
+func metrics() gin.HandlerFunc {
+	h := promhttp.Handler()
+	return func(c *gin.Context) {
+		h.ServeHTTP(c.Writer, c.Request)
+	}
+}
